tty/serial/omap: Make the locking RT aware
The port lock is a spinlock_t which is becomes a sleeping lock on PREEMPT_RT.
The driver splits the locking function into two parts: local_irq_save() and
uart_port_lock(). This helps keeping interrupts disabled in the sysrq case and
acquiring the lock in the other case.

There is no need for this optimisation nor ist there a requirement to keep
interrupts disabled during printing without the lock. In the sysrq and
oops_in_progress case the printing is an best effort attempt 

sysrq case while acquiring the lock in the other cases.

This is not an optimisation we need and it breaks PREEMPT_RT. There is nothing
wrong 

The O pulling out the "_irqsave" from the locking function in order to
The `optimisation' to pull out the "_irqsave" from the locking function and
avoiding the trylock in the sysrq case does not 
sleeping lock and local_irq_save() is not the
optimsation we are looking for. Redo it to make it work on -RT and
non-RT.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>