hrtimer: Allow raw wakeups during boot
There are a few wake-up timers during the early boot which are essencial for
the system to make progress. At this stage there are no softirq spawn for the
softirq processing so there is no timer processing in softirq.
The wakeup in question:

smpboot_create_thread()
 -> kthread_create_on_cpu()
    -> kthread_bind()
       -> wait_task_inactive()
         -> schedule_hrtimeout()

Let the timer fire in hardirq context during the system boot.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>