console: Hide console printk counter for !CONFIG_PRINTK
Each console has its own counter that tracks which printk record it
will print next. However, when !CONFIG_PRINTK, there are no records
to print and there is no need for any such tracking to occur.

Note that this also fixes a build error because some of the functions
necessary for tracking were not exposed with !CONFIG_PRINTK.

Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>